[BUGFIX] Check whether realpath exists in runTests.sh
#94612 introduced the realpath command for retrieving
the "CORE_ROOT" path. This however leads to execution
failures on MacOS systems, which did not manually install
this command (as it's not installed by default).

To prevent the script from failing on default MacOS systems,
a check for the existence of the realpath command is added.
If not installed, the previous behaviour is used while displaying
a short information.

Resolves: #94635
Related: #94612
Releases: master, 10.4, 9.5
Change-Id: I30792f1e5492b57adf7ff28a7fa2c415ac2e094c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70132
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>