[BUGFIX] Streamline mount page resolving for HMENU special=list
TYPO3 used RelationHandler for HMENU special=list to
fetch all pages for this menu once, but fetched
the full records again for MountPoint relation, and
manual language + workspace overlay was taken care later-on.

Once menus were generated with mountpoints and workspace and language
overlays, the ordering was not in line, and existing
queries were not used.

Instead, the code is now simplified and fully relies on
PageRepository functionality, just as most other
Frontend-related code.

Resolves: #94643
Releases: master
Change-Id: Ib86116a74c7db6739c474e5efb13b77284e41b08
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70112
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>