[TASK] Deprecate wincache cache backend
The upstream wincache PHP module is unmaintained, a
PHP 7.4 version came late and there is no work for
a PHP 8 compatible version on the horizon.

The ApcuBackend is a suitable alternative that works
on windows platform as ram driven key/value store.

The patch deprecates the wincache based cache
backend implementation.

Resolves: #94665
Releases: master
Change-Id: I6e1eb551c048f73d48c85ed498d1cd4caab266d3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70130
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>