[TASK] Move a Changelog file to correct directory
The patch for issue #91123 was originally planned for
having a v10 backport, but made it to master only and
has been released as 11.0.
The Important.rst file has been added to 10.4.x
directory, but it should reside in 11.0 directory.
Adapt that and fix a reference to it.

Change-Id: Ia28b3b7d75b27b5f2d1ccd2fe9ec77a767eaeeae
Resolves: #94667
Related: #91123
Releases: master
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70153
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>