[FEATURE] Add filter to columns selector in recordlist
To ease the use of the new columns selector,
especially when dealing with tables, having a
large number of columns (e.g. pages or tt_content),
a filter is added to the columns selector action bar.

The remaining actions, such as "check all" are
then bound to the current filter result. This allows
to quickly select a specific group of columns.

Resolves: #94680
Releases: master
Change-Id: I348fcb9b043a2dd5697248be7aebd322c25cdf3f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70169
Tested-by: core-ci <typo3@b13.com>
Tested-by: Guido Schmechel <guido.schmechel@brandung.de>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>