[BUGFIX] Ensure correct merge of update suggestions
The calls to the getUpdateSuggestions() method may result
in the same change is added twice to the same hash. Using
array_replace_recursive() ensures the same two suggestions
are merged correctly and not as array of strings like done
by the former implementation.

This is important for sqlite which creates multiple
statements for one 'job' sometimes.

Resolves: #94709
Releases: master
Change-Id: I1cec5977215559b17004d0f32b01fec039f09992
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70201
Tested-by: Daniel Siepmann <coding@daniel-siepmann.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Daniel Siepmann <coding@daniel-siepmann.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>