[BUGFIX] Only minifyJS once in Frontend
If you have inline JS (or inlineFooterJS) code
the method "GeneralUtility::minifyJavaScript()"
is called twice for the same code to actually minify
the code, because PageRenderer takes
care of minification, so FE process does not
explicitly have to do that.

This does no harm, only if there is a hook
within minifyJavaScript(), the processing is triggered
twice. This also only happens if config.compressJs = 1
is called.

Resolves: #94716
Releases: master
Change-Id: I77af009bf252b9f249794ed5b8db1dd0ae26b769
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70207
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>