[TASK] Run Grunt scripts in parallel
This patch groups Grunt tasks that can run simultaneously to improve the
build performance. To achieve this, the package `grunt-concurrent` is
used.

On a 2018 Dell XPS 15 with Intel i7-9750H this saves roughly 12 seconds.
To check the difference, the command `time grunt build` was executed
twice:

Sequential build:

  48,09s user 10,06s system 154% cpu 37,565 total
  50,60s user 11,89s system 154% cpu 40,539 total

Parallel build:

  77,91s user 5,96s system 261% cpu 28,311 total
  67,43s user 5,05s system 262% cpu 27,619 total

Due to issues with `grunt-contrib-imagemin`, the subscript `imagemin` is
excluded from the `minify_assets` group for now.

Resolves: #95016
Releases: master
Change-Id: Ic12ff7c0cb43cd0ab050e3e7d1a3a783038189b2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70244
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jonas Eberle <flightvision@googlemail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>