[TASK] Remove docBlockChecker integrity test
We have the annotationChecker and the ever growing
phpstan ruleset that scan our codebase for code and
documentation code block violations.

For the additional script docBlockChecker.php it's
unclear what it actually does and when it should fail.
It never went red to my knowledge. The patch author
didn't know what was going on and I've been unable to
make the script fail with local tests. The mentioned
phpdocumentor/phpdocumentor seems to have never
materialized in our world, and tools like api.typo3.org
rely on doxygen.

The patch drops that integrity test script and removes
usages in runTests.sh and gitlab-ci.

Resolves: #94735
Related: #89023
Releases: master
Change-Id: Id7744780cd50e2009d59bd0f040799f8081924e7
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70277
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>