[BUGFIX] Only allow to edit metadata in FileList if metadata exists
The pencil icon on the file list should only show up if
a metadata record exists, in order to avoid an error.

This case happens if a file wasn't indexed correctly
and the Indexer cannot open the file to create metadata
via the Scheduler, e.g. if the folder of the file has
an invalid character, such as "(blabla)" as suffix and
was uploaded externally via FTP.

Resolves: #94743
Releases: 10.4
Change-Id: I99e23cefc82591fa0562072d61bcc74f72b49b6f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70284
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>