[TASK] Inform about header field usage for special elements
Some content elements, namely "html", "div" and "shortcut"
do not display the header field value in the frontend. However,
the "name" is shown as link value in menus (e.g. section index).

To make this clear to the editor, a new description is added
to the "header" field for those elements. Furthermore is the
corresponding CSH text adjusted.

Resolves: #94704
Releases: master
Change-Id: I9767e41a082e4c2776f0c78e7e091bff9eed66bf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70352
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>