[BUGFIX] Allow shortcut pages to pages of other sites
TYPO3 has had a long history of linking to
other pages in other pagetrees. Most of the issues
were solved with the TYPO3 v9 Routing and Site Handling.

Through this feature, it is now possible to
easily create a page of type shortcut and link to
a page in a different page tree.

This was previously not possible because TYPO3 Core
by default generated URLs like "/index.php?id=123"
where people could simply add ?id=355 (where page 355
resides on a different page tree).

Since this change has been resolved, links to
other pages can be allowed and are automatically
redirected.

If we ever get rid of allowing to call pages with
an `?id=123` query parameter, the $pageNotFound=3 case
can be removed completely.

Resolves: #92859
Resolves: #92750
Releases: main, 11.5
Change-Id: I16ccc3e5b0ccb1419ccd2a8d78443616e7627a33
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70432
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>