[BUGFIX] Rename class for vertical rule in access module
Bootstrap 5.1.0, we updated to in #94752, introduced
a `.vr` (vertical rule) class. This collided with our
access module related styles, since we also use a `.vr`
class for the legend here. To resolve the collision, our
own implementation, which is only used in the access module,
is now prefixed with `t3`.

Resolves: #94826
Related: #94752
Releases: master
Change-Id: I493baa3d0af41b78f42c9e73c9c121b29f417edd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70456
Tested-by: core-ci <typo3@b13.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>