[BUGFIX] Do not transform text when RTE is disabled
The TcaText data provider transforms a given database
field value with the RteHtmlParser as soon as `enableRichtext`
is set for the field in TCA. This means, the data provider
did previously not checked, whether RTE is disabled for the
user in general or for this specific field by configuration
(e.g. page TSconfig). This therefore led, among others, to
"normal" text being falsely wrapped with `<p>` tags.

This is now fixed by checking all relevant settings before
actually transforming the database field value.

Resolves: #94915
Releases: master, 10.4
Change-Id: I5af4e9997c28269be959ea9e63e12cfee5e991af
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70737
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>