[BUGFIX] Re-enable reset of excluded records in EXT:impexp
This is a regression introduced with #94057 that causes resetting
excluded records to throw an exception in the export view, since the
incoming parameter "exclude" could be of type array (= add excluded
records) or integer (= include all records again), but was only
checked for null or not null, which was ambiguous in the second case.
For clear separation the reset of excluded records is now done via a
new parameter "resetExclude".

Resolves: #94949
Releases: master
Change-Id: Iddb8d21e0f75cc24ef18dd2f851a479cb7dafb3b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70690
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Alexander Nitsche <typo3@alexandernitsche.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Alexander Nitsche <typo3@alexandernitsche.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>