[BUGFIX] Properly check if workspace is accessible
The workspace JavaScript component tries to
fetch the workspace listing on accessing the
backend module. Since #94819 this results in
a JavaScript TypeError since the provided selector
for the progress bar component, used as indicator
for the ajax request, might not exist.

This is now fixed by checking if the the workspace
is actually accessible prior to initiating the ajax call.

Resolves: #94946
Related: #94819
Releases: master
Change-Id: I0c77564acc197a57e2999f4439daf28ba3fd333c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70724
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>