[FEATURE] Implement multi record selection in recodlist
The multi record selection component, introduced
in #94906, is now added to the recordlist as well.

Since the component is not bound to the clipboard
functionality, this will solve a lot of issues,
editors faced when trying to perform actions for
multiple records at once. For example, it was
previously required to activate the clipboard,
change to a numeric "pad" and to copy records
to this pad, before any action could be performed.

This patch also resolves some smaller bugs related
to clipboard actions and does some code cleanup in
the DatabaseRecordList class as well as in the
corresponding JavaScript module.

Resolves: #95068
Related: #94906
Releases: master
Change-Id: I5bd3120f71cd4043d778f8116b3cf643b6e72064
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70803
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Oliver Bartsch <bo@cedev.de>