[TASK] Streamline composer.json deficit detector
Code for scanning extension folders and populating
information from that, is removed to use already
available API in extension manager.

Therefore the ComposerDeficitDetector, which has
only usages in EM anyway, is moved into the namespace
of EM.

With doing so some folders in typo3conf/ext might be
missed, that do not contain valid extensions,
but for these, providing composer.json suggetions
might not be useful anyway.

Releases: master
Resolves: #95063
Related: #93931
Change-Id: I7652036a7c0a4ac9e9884393383945f58788d88e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70850
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benni Mack <benni@typo3.org>