[TASK] Use existing site instance in AbstractItemProvider
Since #94662 it's possible to access the site
configuration in `foreign_table_where` clauses.

The corresponding site configuration was therefore
fetched using the SiteFinder, which is however not
needed as the SiteResolving data provider already
does this.

This path therefore replaces the duplicated call to
SiteFinder by accessing the Site instance, already
present in the result array.

Resolves: #95033
Related: #94662
Releases: master
Change-Id: I6ea8f09721ff6a3963c1521b8b1476d7e39ddcaf
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70821
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>