[TASK] Move table editor the ce edit view
The table editor is currently accessed by using
a seperated wizard next to the bodytext field.
On each and every "edit" a user needs to click
this button to get the table editor.

This has been changed by moving the table
editor to the regular content edit view and
replaces the text area with the table editor.

The fields 'table_delimiter' and 'table_enclosure'
are not working as expected, but this is not touched
by this patch but will hopefully be taken care of
in the future.

Resolves: #95036
Releases: master
Change-Id: I9497987f9d64774b2d1272e4c269ef766caf760c
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70823
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>