[TASK] Do not handle component specific logic in multi record selection
The multi record selection component is responsible
for managing the selection of records as well as
toggling the corresponding actions. It should not
implement any component specific action logic,
but just dispatch a custom event, informing the
component about the requested action.

Therefore, the "edit file metadata" handling is
moved into the corresponding FileList component.

To make life easier for implementing components, the
details, any dispatched action event contains, are now
extended for the action specific configuration.

Furthermore are the ActionConfiguration and
ActionEventDetails interfaces now properly exported.
This allows to increase type safety in the implementing
components.

Resolves: #95040
Releases: master
Change-Id: I1a5da3eacee5ee86ae5cbacbad8f795c41674c80
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70825
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>