[BUGFIX] Skip localize option in language switch
The language switch in FormEngine allows
to either switch to an existing translation
of the current record, or to create such a
translation. This however currently does only
work in case a default language record exists.

Therefore, in case no default record exists,
the "localize" option is skipped for the record,
preventing PHP warnings.

Resolves: #95160
Releases: master
Change-Id: I01211b16fc51c8a18a7f97f5d5c8fa316a34ae49
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70979
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benni Mack <benni@typo3.org>