[TASK] Use executeFrontendSubRequest() for SecureHtmlRenderingTest
Executing frontend request calls in tests has been transformed to
the new subrequest support of v11.

One test has been added with old executeFrontendRequest() which is
deprecated in the testing-framework and is changed to the new
executeFrontendSubRequest() ability.

This patch acts as post-patch to change this for master.

Resolves: #95178
Releases: master
Change-Id: I773b0131d6bec1e2191cd861f9c65db0fff0c81a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71017
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>